{common}{ros2} Add ROS opt prefix to PYTHONPATH
In order for Python to find modules in the recipe-sysroot
we must add the path to PYTHONPATH.  However, we must prepend
the path to any previous values of PYTHONPATH.

We cannot refer to the variable in its own definition since
bitbake does not evaluate strings in that way.

Instead we must mimick python3targetconfig.bbclass and
prepend to the recipe tasks to conditionally set PYTHONPATH
in the shell context.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>